feature: Add increment command [#24]
Add a safe way to increment counters with BoJack.

Our "users" should not have to deal with concurrency when trying to create a safe counter with BoJack.

e.g.:

counter = client.get("counter")

if counter
  counter = counter.to_i
  counter += 1
  client.set("counter", counter)
end

That is completely unsafe if you have concurrent access, imagine this in a multi-threaded web-server.

For this I purpose the command increment $key so the users can rely upon us to increment the key value by 1.

Important:

Only valid keys can be incremented, to be valid the key MUST already exist and be MUST be able to be casted to Integer, meaning that it cannot be an Array or a non-numerical String.